+++ src/App.jsx
@@ -3,7 +3,9 @@ import AllPostsPage from "./pages/AllPosts";
 import NewPostPage from "./pages/NewPosts";
 import FavouritesPage from "./pages/Favourites";
 import MainNavigation from "./components/MainNavigation";
-import Login from './Login';
+import Login from './auth/Login';
+import ProtectedRoute from "./components/protectedRoute";
+import Dashboard from "./Dashboard";
 
 
 
@@ -12,11 +14,13 @@ const App = () => {
     <>
       <MainNavigation />
       <Switch>
-        <Route exact path="/" component={Login} />
+        <Route exact path="/" component={ProtectedRoute}>
+          <Route exact path="/" component={Dashboard} />
+        </Route>
       </Switch>
 
       <Switch>
-        <Route path="/" exact>
+        <Route path="/dashboard" exact>
           <AllPostsPage />
         </Route>
         <Route path="/new-post">
