--- src/firebase.js
@@ -1,90 +1,0 @@
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-import {GoogleAuthProvider, getAuth, signInWithPopup, signInWithEmailAndPassword, createUserWithEmailAndPassword, sendPasswordResetEmail,signOut,} from "firebase/auth";
-
-import {getFirestore, query, getDocs, collection, where, addDoc} from "firebase/firestore"; 
-
-const firebaseConfig = {
-    apiKey: "AIzaSyD7wtD7eFqK-KKkaZYvqYX2q3UumS5cR2k",
-    authDomain: "community-hub-4.firebaseapp.com",
-    projectId: "community-hub-4",
-    storageBucket: "community-hub-4.appspot.com",
-    messagingSenderId: "1058932236902",
-    appId: "1:1058932236902:web:fd79613b89797970ca03a7",
-    measurementId: "G-NTNNQ4DNZQ"
-  };
-
-const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
-const db = getFirestore(app);
-
-const googleProvider = new GoogleAuthProvider();
-const signInWithGoogle = async () => {
-  try {
-    const res = await signInWithPopup(auth, googleProvider);
-    const user = res.user;
-    const q = query(collection(db, "users"), where("uid", "==", user.uid));
-    const docs = await getDocs(q);
-    if (docs.docs.length === 0) {
-      await addDoc(collection(db, "users"), {
-        uid: user.uid,
-        name: user.displayName,
-        authProvider: "google",
-        email: user.email,
-      });
-    }
-  } catch (err) {
-    console.error(err);
-    alert(err.message);
-  }
-};
-
-const logInWithEmailAndPassword = async (email, password) => {
-    try {
-      await signInWithEmailAndPassword(auth, email, password);
-    } catch (err) {
-      console.error(err);
-      alert(err.message);
-    }
-  };
-
-  const registerWithEmailAndPassword = async (name, email, password) => {
-    try {
-      const res = await createUserWithEmailAndPassword(auth, email, password);
-      const user = res.user;
-      await addDoc(collection(db, "users"), {
-        uid: user.uid,
-        name,
-        authProvider: "local",
-        email,
-      });
-    } catch (err) {
-      console.error(err);
-      alert(err.message);
-    }
-  };
-
-  const sendPasswordReset = async (email) => {
-    try {
-      await sendPasswordResetEmail(auth, email);
-      alert("Password reset link sent!");
-    } catch (err) {
-      console.error(err);
-      alert(err.message);
-    }
-  };
-
-  const logout = () => {
-    signOut(auth);
-  };
-
-  export {
-    auth,
-    db,
-    signInWithGoogle,
-    signInWithEmailAndPassword,
-    logInWithEmailAndPassword,
-    registerWithEmailAndPassword,
-    sendPasswordReset,
-    logout,
-  };
